chore: refactor util
