--- apps/server/internal/util/account.go
@@ -1,22 +1,0 @@
-package util
-
-//
-// import (
-// 	"fijoy/internal/gen/postgres/model"
-//
-// 	fijoyv1 "fijoy/internal/gen/proto/fijoy/v1"
-// )
-//
-// var JetAccountTypeToConnectAccountType = map[model.FijoyAccountType]fijoyv1.AccountType{
-// 	model.FijoyAccountType_Cash:       fijoyv1.AccountType_ACCOUNT_TYPE_CASH,
-// 	model.FijoyAccountType_Debt:       fijoyv1.AccountType_ACCOUNT_TYPE_DEBT,
-// 	model.FijoyAccountType_Investment: fijoyv1.AccountType_ACCOUNT_TYPE_INVESTMENT,
-// 	model.FijoyAccountType_OtherAsset: fijoyv1.AccountType_ACCOUNT_TYPE_OTHER_ASSET,
-// }
-//
-// var ConnectAccountTypeToJetAccountType = map[fijoyv1.AccountType]model.FijoyAccountType{
-// 	fijoyv1.AccountType_ACCOUNT_TYPE_CASH:        model.FijoyAccountType_Cash,
-// 	fijoyv1.AccountType_ACCOUNT_TYPE_DEBT:        model.FijoyAccountType_Debt,
-// 	fijoyv1.AccountType_ACCOUNT_TYPE_INVESTMENT:  model.FijoyAccountType_Investment,
-// 	fijoyv1.AccountType_ACCOUNT_TYPE_OTHER_ASSET: model.FijoyAccountType_OtherAsset,
-// }
