+++ apps/server/internal/util/auth/auth.go
@@ -1,4 +1,4 @@
-package util
+package auth
 
 import (
 	"context"
