--- apps/server/internal/util/money_test.go
@@ -1,6 +1,0 @@
-package util
-
-import "testing"
-
-func TestSomething(t *testing.T) {
-}
